refactor: extract grid selection column base mixin (#6143)
* refactor: extract GridSelectionColumnBaseMixin

* Update docs and Typescript definitions

* Remove mixes annotation from TS definition

* Add protected methods to TS definition
