+++ packages/grid/src/vaadin-grid-selection-column.js
@@ -4,8 +4,8 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import '@vaadin/checkbox/src/vaadin-checkbox.js';
-import { addListener } from '@vaadin/component-base/src/gestures.js';
 import { GridColumn } from './vaadin-grid-column.js';
+import { GridSelectionColumnBaseMixin } from './vaadin-grid-selection-column-base-mixin.js';
 
 /**
  * `<vaadin-grid-selection-column>` is a helper element for the `<vaadin-grid>`
@@ -29,9 +29,10 @@ import { GridColumn } from './vaadin-grid-column.js';
  *
  * __The default content can also be overridden__
  *
+ * @mixes GridSelectionColumnBaseMixin
  * @fires {CustomEvent} select-all-changed - Fired when the `selectAll` property changes.
  */
-class GridSelectionColumn extends GridColumn {
+class GridSelectionColumn extends GridSelectionColumnBaseMixin(GridColumn) {
   static get is() {
     return 'vaadin-grid-selection-column';
   }
@@ -39,74 +40,16 @@ class GridSelectionColumn extends GridColumn {
   static get properties() {
     return {
       /**
-       * Width of the cells for this column.
-       */
-      width: {
-        type: String,
-        value: '58px',
-      },
-
-      /**
-       * Flex grow ratio for the cell widths. When set to 0, cell width is fixed.
-       * @attr {number} flex-grow
-       * @type {number}
-       */
-      flexGrow: {
-        type: Number,
-        value: 0,
-      },
-
-      /**
-       * When true, all the items are selected.
-       * @attr {boolean} select-all
-       * @type {boolean}
-       */
-      selectAll: {
-        type: Boolean,
-        value: false,
-        notify: true,
-      },
-
-      /**
-       * When true, the active gets automatically selected.
-       * @attr {boolean} auto-select
-       * @type {boolean}
-       */
-      autoSelect: {
-        type: Boolean,
-        value: false,
-      },
-
-      /**
-       * When true, rows can be selected by dragging over the selection column.
-       * @attr {boolean} drag-select
-       * @type {boolean}
-       */
-      dragSelect: {
-        type: Boolean,
-        value: false,
-      },
-
-      /** @private */
-      __indeterminate: Boolean,
-
-      /**
        * The previous state of activeItem. When activeItem turns to `null`,
        * previousActiveItem will have an Object with just unselected activeItem
        * @private
        */
       __previousActiveItem: Object,
-
-      /** @private */
-      __selectAllHidden: Boolean,
     };
   }
 
   static get observers() {
-    return [
-      '__onSelectAllChanged(selectAll)',
-      '_onHeaderRendererOrBindingChanged(_headerRenderer, _headerCell, path, header, selectAll, __indeterminate, __selectAllHidden)',
-    ];
+    return ['__onSelectAllChanged(selectAll)'];
   }
 
   constructor() {
@@ -136,50 +79,6 @@ class GridSelectionColumn extends GridColumn {
       this._grid.addEventListener('filter-changed', this.__boundOnSelectedItemsChanged);
       this._grid.addEventListener('selected-items-changed', this.__boundOnSelectedItemsChanged);
     }
-  }
-
-  /**
-   * Renders the Select All checkbox to the header cell.
-   *
-   * @override
-   */
-  _defaultHeaderRenderer(root, _column) {
-    let checkbox = root.firstElementChild;
-    if (!checkbox) {
-      checkbox = document.createElement('vaadin-checkbox');
-      checkbox.setAttribute('aria-label', 'Select All');
-      checkbox.classList.add('vaadin-grid-select-all-checkbox');
-      checkbox.addEventListener('checked-changed', this.__onSelectAllCheckedChanged.bind(this));
-      root.appendChild(checkbox);
-    }
-
-    const checked = this.__isChecked(this.selectAll, this.__indeterminate);
-    checkbox.__rendererChecked = checked;
-    checkbox.checked = checked;
-    checkbox.hidden = this.__selectAllHidden;
-    checkbox.indeterminate = this.__indeterminate;
-  }
-
-  /**
-   * Renders the Select Row checkbox to the body cell.
-   *
-   * @override
-   */
-  _defaultRenderer(root, _column, { item, selected }) {
-    let checkbox = root.firstElementChild;
-    if (!checkbox) {
-      checkbox = document.createElement('vaadin-checkbox');
-      checkbox.setAttribute('aria-label', 'Select Row');
-      checkbox.addEventListener('checked-changed', this.__onSelectRowCheckedChanged.bind(this));
-      root.appendChild(checkbox);
-      addListener(root, 'track', this.__onCellTrack.bind(this));
-      root.addEventListener('mousedown', this.__onCellMouseDown.bind(this));
-      root.addEventListener('click', this.__onCellClick.bind(this));
-    }
-
-    checkbox.__item = item;
-    checkbox.__rendererChecked = selected;
-    checkbox.checked = selected;
   }
 
   /** @private */
@@ -217,182 +116,49 @@ class GridSelectionColumn extends GridColumn {
   }
 
   /**
-   * Enables or disables the Select All mode once the Select All checkbox is switched.
-   * The listener handles only user-fired events.
+   * Override a method from `GridSelectionColumnBaseMixin` to handle the user
+   * selecting all items.
    *
-   * @private
+   * @protected
+   * @override
    */
-  __onSelectAllCheckedChanged(e) {
-    // Skip if the state is changed by the renderer.
-    if (e.target.checked === e.target.__rendererChecked) {
-      return;
-    }
-
-    this.selectAll = this.__indeterminate || e.target.checked;
+  _selectAll() {
+    this.selectAll = true;
   }
 
   /**
-   * Selects or deselects the row once the Select Row checkbox is switched.
-   * The listener handles only user-fired events.
+   * Override a method from `GridSelectionColumnBaseMixin` to handle the user
+   * deselecting all items.
    *
-   * @private
+   * @protected
+   * @override
    */
-  __onSelectRowCheckedChanged(e) {
-    // Skip if the state is changed by the renderer.
-    if (e.target.checked === e.target.__rendererChecked) {
-      return;
-    }
-
-    if (e.target.checked) {
-      this._grid.selectItem(e.target.__item);
-    } else {
-      this._grid.deselectItem(e.target.__item);
-    }
-  }
-
-  /** @private */
-  __onCellTrack(event) {
-    if (!this.dragSelect) {
-      return;
-    }
-    this.__dragCurrentY = event.detail.y;
-    this.__dragDy = event.detail.dy;
-    if (event.detail.state === 'start') {
-      const visibleRows = this._grid._getVisibleRows();
-      // Get the row where the drag started
-      const dragStartRow = visibleRows.find((row) => row.contains(event.currentTarget.assignedSlot));
-      // Whether to select or deselect the items on drag
-      this.__dragSelect = !this._grid._isSelected(dragStartRow._item);
-      // Store the index of the row where the drag started
-      this.__dragStartIndex = dragStartRow.index;
-      // Store the item of the row where the drag started
-      this.__dragStartItem = dragStartRow._item;
-      // Start the auto scroller
-      this.__dragAutoScroller();
-    } else if (event.detail.state === 'end') {
-      // If drag start and end stays within the same item, then toggle its state
-      if (this.__dragStartItem) {
-        if (this.__dragSelect) {
-          this._grid.selectItem(this.__dragStartItem);
-        } else {
-          this._grid.deselectItem(this.__dragStartItem);
-        }
-      }
-      // Clear drag state after timeout, which allows preventing the
-      // subsequent click event if drag started and ended on the same item
-      setTimeout(() => {
-        this.__dragStartIndex = undefined;
-      });
-    }
-  }
-
-  /** @private */
-  __onCellMouseDown(e) {
-    if (this.dragSelect) {
-      // Prevent text selection when starting to drag
-      e.preventDefault();
-    }
-  }
-
-  /** @private */
-  __onCellClick(e) {
-    if (this.__dragStartIndex !== undefined) {
-      // Stop the click event if drag was enabled. This click event should
-      // only occur if drag started and stopped on the same item. In that case
-      // the selection state has already been toggled on drag end, and we
-      // don't  want to toggle it again from clicking the checkbox or changing
-      // the active item.
-      e.preventDefault();
-    }
-  }
-
-  /** @private */
-  __dragAutoScroller() {
-    if (this.__dragStartIndex === undefined) {
-      return;
-    }
-    // Get the row being hovered over
-    const visibleRows = this._grid._getVisibleRows();
-    const hoveredRow = visibleRows.find((row) => {
-      const rowRect = row.getBoundingClientRect();
-      return this.__dragCurrentY >= rowRect.top && this.__dragCurrentY <= rowRect.bottom;
-    });
-
-    // Get the index of the row being hovered over or the first/last
-    // visible row if hovering outside the grid
-    let hoveredIndex = hoveredRow ? hoveredRow.index : undefined;
-    const scrollableArea = this.__getScrollableArea();
-    if (this.__dragCurrentY < scrollableArea.top) {
-      hoveredIndex = this._grid._firstVisibleIndex;
-    } else if (this.__dragCurrentY > scrollableArea.bottom) {
-      hoveredIndex = this._grid._lastVisibleIndex;
-    }
-
-    if (hoveredIndex !== undefined) {
-      // Select all items between the start and the current row
-      visibleRows.forEach((row) => {
-        if (
-          (hoveredIndex > this.__dragStartIndex && row.index >= this.__dragStartIndex && row.index <= hoveredIndex) ||
-          (hoveredIndex < this.__dragStartIndex && row.index <= this.__dragStartIndex && row.index >= hoveredIndex)
-        ) {
-          if (this.__dragSelect) {
-            this._grid.selectItem(row._item);
-          } else {
-            this._grid.deselectItem(row._item);
-          }
-          this.__dragStartItem = undefined;
-        }
-      });
-    }
-
-    // Start scrolling in the top/bottom 15% of the scrollable area
-    const scrollTriggerArea = scrollableArea.height * 0.15;
-    // Maximum number of pixels to scroll per iteration
-    const maxScrollAmount = 10;
-
-    if (this.__dragDy < 0 && this.__dragCurrentY < scrollableArea.top + scrollTriggerArea) {
-      const dy = scrollableArea.top + scrollTriggerArea - this.__dragCurrentY;
-      const percentage = Math.min(1, dy / scrollTriggerArea);
-      this._grid.$.table.scrollTop -= percentage * maxScrollAmount;
-    }
-    if (this.__dragDy > 0 && this.__dragCurrentY > scrollableArea.bottom - scrollTriggerArea) {
-      const dy = this.__dragCurrentY - (scrollableArea.bottom - scrollTriggerArea);
-      const percentage = Math.min(1, dy / scrollTriggerArea);
-      this._grid.$.table.scrollTop += percentage * maxScrollAmount;
-    }
-
-    // Schedule the next auto scroll
-    setTimeout(() => this.__dragAutoScroller(), 10);
+  _deselectAll() {
+    this.selectAll = false;
   }
 
   /**
-   * Gets the scrollable area of the grid as a bounding client rect. The
-   * scrollable area is the bounding rect of the grid minus the header and
-   * footer.
+   * Override a method from `GridSelectionColumnBaseMixin` to handle the user
+   * selecting an item.
    *
-   * @private
+   * @param {Object} item the item to select
+   * @protected
+   * @override
    */
-  __getScrollableArea() {
-    const gridRect = this._grid.$.table.getBoundingClientRect();
-    const headerRect = this._grid.$.header.getBoundingClientRect();
-    const footerRect = this._grid.$.footer.getBoundingClientRect();
-
-    return {
-      top: gridRect.top + headerRect.height,
-      bottom: gridRect.bottom - footerRect.height,
-      left: gridRect.left,
-      right: gridRect.right,
-      height: gridRect.height - headerRect.height - footerRect.height,
-      width: gridRect.width,
-    };
+  _selectItem(item) {
+    this._grid.selectItem(item);
   }
 
   /**
-   * IOS needs indeterminate + checked at the same time
-   * @private
+   * Override a method from `GridSelectionColumnBaseMixin` to handle the user
+   * deselecting an item.
+   *
+   * @param {Object} item the item to deselect
+   * @protected
+   * @override
    */
-  __isChecked(selectAll, indeterminate) {
-    return indeterminate || selectAll;
+  _deselectItem(item) {
+    this._grid.deselectItem(item);
   }
 
   /** @private */
@@ -419,13 +185,13 @@ class GridSelectionColumn extends GridColumn {
       this.__withFilteredItemsArray((items) => {
         if (!this._grid.selectedItems.length) {
           this.selectAll = false;
-          this.__indeterminate = false;
+          this._indeterminate = false;
         } else if (this.__arrayContains(this._grid.selectedItems, items)) {
           this.selectAll = true;
-          this.__indeterminate = false;
+          this._indeterminate = false;
         } else {
           this.selectAll = false;
-          this.__indeterminate = true;
+          this._indeterminate = true;
         }
       });
     }
@@ -434,7 +200,7 @@ class GridSelectionColumn extends GridColumn {
 
   /** @private */
   __onDataProviderChanged() {
-    this.__selectAllHidden = !Array.isArray(this._grid.items);
+    this._selectAllHidden = !Array.isArray(this._grid.items);
   }
 
   /**
